Critter Overhaul
Have made a blueprint for a new kind of critter that gives gameplay a bit more dynamism.

Spawned Bones spawn BP_Beetle

BP_Beetle flies about the bone that spawned it

When it hears a hoot, it emits an effect making it easier to see

Can be clicked on, causing the owl to dive toward it (still buggy)

If the owl touches it when diving, it will be eaten (still buggy)

Still some polishing to do!